GCC-7: disable pre-C++17 warning
This warning is only relevant when object files compile with and without
C++17 language level are to be linked into a single executable; starting
with C++17, new style 'noexcept' specifications will become part of the
function signature and thus part of the mangled function name. Linkiing
mixed object files might fail in such a situation.

Obviously this warning is not relevant for us; moreover we plan to
upgrade to C++17 soon