ChainSearch: now use the improved TreeExplorer capabilities to address the shortcoming
...the solution built thus far was logically broken, since it retained the unfiltered
source sequence within the base layer. Thus it would backtrack into this unfiltered
sequence eventually.

The idea was to build a special treatment for attaching the first filter condition;
in fact the first one does not need to be added to the chain, but rather should be
planted directly into the base layer.

WIP: this is a solution draft, but does not work yet
  - when attaching the base layer, the filter is pulled twice
  - an overconstrained filter raises an Assertion failure
    (instead of just transitioning into exhausted state)